login creds and authentication added

+++ nursery_management/src/main/java/com/nursery/management/controller/AdminController.java
@@ -6,6 +6,10 @@ import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.cache.annotation.CacheEvict;
 import org.springframework.http.HttpStatus;
 import org.springframework.http.ResponseEntity;
+import org.springframework.security.access.prepost.PreAuthorize;
+import org.springframework.security.core.Authentication;
+import org.springframework.security.core.context.SecurityContextHolder;
+import org.springframework.security.core.userdetails.UserDetails;
 import org.springframework.web.bind.annotation.CrossOrigin;
 import org.springframework.web.bind.annotation.DeleteMapping;
 import org.springframework.web.bind.annotation.GetMapping;
@@ -17,6 +21,7 @@ import org.springframework.web.bind.annotation.RequestMapping;
 import org.springframework.web.bind.annotation.RestController;
 
 import com.nursery.management.entity.Admin;
+import com.nursery.management.entity.CareTaker;
 import com.nursery.management.entity.Nursery;
 import com.nursery.management.service.AdminService;
 import com.nursery.management.service.NurseryService;
@@ -31,30 +36,73 @@ public class AdminController {
 	@Autowired
 	private AdminService adminService;
 
-	@Autowired
-	private NurseryService nurseryService;
-
 	@GetMapping
+	@PreAuthorize("hasRole('ROLE_SUPER_ADMIN')")
 	public List<Admin> getAllAdmins() {
 		return adminService.getAllAdmins();
 	}
 
 	@GetMapping("/byId/{adminId}")
-	public Admin getAdminById(@PathVariable String adminId) {
-		return adminService.getAdminById(adminId);
+	@PreAuthorize("hasRole('ROLE_ADMIN') or hasRole('ROLE_SUPER_ADMIN')")
+	public ResponseEntity<?> getAdminById(@PathVariable String adminId) {
+		Admin admin = adminService.getAdminById(adminId);
+
+		Authentication authentication = SecurityContextHolder.getContext().getAuthentication();
+
+		if (authentication != null && authentication.isAuthenticated()) {
+			Object principal = authentication.getPrincipal();
+			String nursery_id = "";
+			String role = "";
+			if (principal instanceof UserDetails) {
+				UserDetails userDetails = (UserDetails) principal;
+				nursery_id = userDetails.getUsername().split(":")[1];
+				role = userDetails.getAuthorities().toArray()[0].toString();
+			}
+			if ((!role.equals("ROLE_SUPER_ADMIN")) && !nursery_id.isBlank() && !nursery_id.isEmpty()
+					&& !nursery_id.equals(admin.getNursery().getNurseryId())) {
+				return ResponseEntity.status(HttpStatus.UNAUTHORIZED).body("You dont have access to this!");
+			}
+		}
+		if (admin != null) {
+			return ResponseEntity.ok(admin);
+		} else {
+			return ResponseEntity.status(HttpStatus.NOT_FOUND).build();
+		}
 
 	}
 
 	@GetMapping("/byNursery/{nurseryId}")
-	public List<Admin> getAdminsByNursery(@PathVariable String nurseryId) {
+	@PreAuthorize("hasRole('ROLE_ADMIN') or hasRole('ROLE_SUPER_ADMIN')")
+	public ResponseEntity<?> getAdminsByNursery(@PathVariable String nurseryId) {
 
-		Nursery nursery = nurseryService.getNurseryById(nurseryId);
-		return adminService.getAdminsByNursery(nursery);
+		List<Admin> admins = adminService.getAdminsByNurseryId(nurseryId);
+
+		Authentication authentication = SecurityContextHolder.getContext().getAuthentication();
+
+		if (authentication != null && authentication.isAuthenticated()) {
+			Object principal = authentication.getPrincipal();
+			String nursery_id = "";
+			String role = "";
+			if (principal instanceof UserDetails) {
+				UserDetails userDetails = (UserDetails) principal;
+				nursery_id = userDetails.getUsername().split(":")[1];
+				role = userDetails.getAuthorities().toArray()[0].toString();
+			}
+			if ((!role.equals("ROLE_SUPER_ADMIN")) && !nursery_id.isBlank() && !nursery_id.isEmpty()
+					&& !nursery_id.equals(nurseryId)) {
+				return ResponseEntity.status(HttpStatus.UNAUTHORIZED).body("You dont have access to this!");
+			}
+		}
+		if (admins != null) {
+			return ResponseEntity.ok(admins);
+		} else {
+			return ResponseEntity.status(HttpStatus.NOT_FOUND).build();
+		}
 
 	}
 
 	@PostMapping("/createAdmin")
-
+	@PreAuthorize("hasRole('ROLE_ADMIN') or hasRole('ROLE_SUPER_ADMIN')")
 	public ResponseEntity<?> createAdmin(@RequestBody Admin admin) {
 		try {
 			Admin createdAdmin = adminService.createAdmin(admin);
@@ -67,23 +115,60 @@ public class AdminController {
 	}
 
 	@PutMapping("/{adminId}")
+	@PreAuthorize("hasRole('ROLE_ADMIN') or hasRole('ROLE_SUPER_ADMIN')")
 	public ResponseEntity<?> updateAdmin(@PathVariable String adminId, @RequestBody Admin updatedAdmin) {
-		try {
-			Admin admin = adminService.updateAdmin(adminId, updatedAdmin);
-			return ResponseEntity.ok(admin);
-		} catch (EntityNotFoundException e) {
-			e.printStackTrace();
-			return ResponseEntity.status(HttpStatus.NOT_FOUND).body(e.getMessage());
+		Admin admin = adminService.getAdminById(adminId);
+		Authentication authentication = SecurityContextHolder.getContext().getAuthentication();
+
+		if (authentication != null && authentication.isAuthenticated()) {
+			Object principal = authentication.getPrincipal();
+			String nursery_id = "";
+			String role = "";
+			if (principal instanceof UserDetails) {
+				UserDetails userDetails = (UserDetails) principal;
+				nursery_id = userDetails.getUsername().split(":")[1];
+				role = userDetails.getAuthorities().toArray()[0].toString();
+			}
+			if ((!role.equals("ROLE_SUPER_ADMIN")) && !nursery_id.isBlank() && !nursery_id.isEmpty()
+					&& !nursery_id.equals(admin.getNursery().getNurseryId())) {
+				return ResponseEntity.status(HttpStatus.UNAUTHORIZED).body("You dont have access to this!");
+			}
+		}
+		if (admin != null) {
+			updatedAdmin.setAdminId(adminId);
+			Admin savedAdmin = adminService.updateAdmin(adminId, updatedAdmin);
+			return ResponseEntity.ok(savedAdmin);
+		} else {
+			return ResponseEntity.status(HttpStatus.NOT_FOUND).build();
 		}
 	}
 
 	@DeleteMapping("/{adminId}")
-	public ResponseEntity<Void> deleteAdmin(@PathVariable String adminId) {
-		try {
+	@PreAuthorize("hasRole('ROLE_ADMIN') or hasRole('ROLE_SUPER_ADMIN')")
+	public ResponseEntity<?> deleteAdmin(@PathVariable String adminId) {
+		Admin admin = adminService.getAdminById(adminId);
+		Authentication authentication = SecurityContextHolder.getContext().getAuthentication();
+
+		if (authentication != null && authentication.isAuthenticated()) {
+			Object principal = authentication.getPrincipal();
+			String nursery_id = "";
+			String role = "";
+			if (principal instanceof UserDetails) {
+				UserDetails userDetails = (UserDetails) principal;
+				nursery_id = userDetails.getUsername().split(":")[1];
+				role = userDetails.getAuthorities().toArray()[0].toString();
+			}
+			if ((!role.equals("ROLE_SUPER_ADMIN")) && !nursery_id.isBlank() && !nursery_id.isEmpty()
+					&& !nursery_id.equals(admin.getNursery().getNurseryId())) {
+				return ResponseEntity.status(HttpStatus.UNAUTHORIZED).body("You dont have access to this!");
+			}
+		}
+
+		if (admin != null) {
 			adminService.deleteAdmin(adminId);
 			return ResponseEntity.noContent().build();
-		} catch (EntityNotFoundException e) {
-			return ResponseEntity.notFound().build();
+		} else {
+			return ResponseEntity.status(HttpStatus.NOT_FOUND).build();
 		}
 	}
 }
